Rework RST frame handling (#15)
some of the stream frame processing logic was moved to the stream
receiver itself now that it's a task.

the main reason is so we can receive all frames before an RST. Before
the change, getting an RST would just close the stream without consuming
all data frames. The alternative is to close the queue gracefully so it
can be consumed until the end before closing, maybe using a poison pill.
But I don't like the dispatcher closing streams anyway.