Improve frame creation (#44)
this shows no improvement btw, maybe on yasync it does

This revealed a race condition in the stream recv transition. Raises
cause race conditions as they are Future.fail() calls that set callbacks
to continue later. In the meantime an END_STREAM may have been sent.
So the RST won't be sent.

But I think this does not matter. The only time RST won't be sent is if
the stream got closed, which gives the same result: the stream is
closed. However, this makes the h2spec fail.